Refactor create recipe components
