+++ src/components/ingredients/IngredientModal.tsx
@@ -2,82 +2,57 @@ import {
   Button,
   ButtonGroup,
   Flex,
-  Heading,
   Modal,
   ModalBody,
   ModalContent,
   ModalOverlay,
-  Text,
 } from "@chakra-ui/react";
+import { FC, RefObject, useRef, useState } from "react";
 import {
-  Dispatch,
-  FC,
-  RefObject,
-  SetStateAction,
-  useEffect,
-  useRef,
-  useState,
-} from "react";
-import CreateIngredientInput from "./CreateIngredientInput.tsx";
-import ChosenProducts from "./ChosenProducts.tsx";
-import { useQuery } from "@tanstack/react-query";
-import { fetchAll } from "../api/api.ts";
-import { useMyStore } from "../store/store.tsx";
-import { SimpleIngredient, NewIngredient, Product } from "../types.ts";
-import Autocomplete from "./Autocomplete.tsx";
+  useMyStore,
+  // usePurgeStorage,
+} from "../store/store.tsx";
+import { SimpleIngredient, NewIngredient } from "../types.ts";
+import IngredientModalOne from "./IngredientModalOne.tsx";
 import IngredientModalTwo from "./IngredientModalTwo.tsx";
-import { useDebounce } from "@uidotdev/usehooks";
 
 type Props = {
   isOpen: boolean;
   onClose: () => void;
   focusRef: RefObject<HTMLInputElement>;
-  selectedIngredients: NewIngredient[];
-  setSelectedIngredients: Dispatch<SetStateAction<NewIngredient[]>>;
 };
 
-const IngredientModal: FC<Props> = ({
-  isOpen,
-  onClose,
-  focusRef,
-  setSelectedIngredients,
-  selectedIngredients,
-}) => {
-  const { addIngredient, ingredients } = useMyStore();
+const IngredientModal: FC<Props> = ({ isOpen, onClose, focusRef }) => {
+  const {
+    addIngredient,
+    addToSelectedIngredients,
+    selectedProducts,
+    selectedIngredient,
+    editSelectedProducts,
+    selectIngredient,
+  } = useMyStore();
   // const purgeStorage = usePurgeStorage();
-  // const bytesInUse = useBytesInUse();
-  const [error, setError] = useState(false);
+  // // const bytesInUse = useBytesInUse();
   const [name, setName] = useState<string | null>(null);
-  const [currentIngredients, setCurrentIngredients] =
-    useState<NewIngredient[]>(ingredients);
   const [amount, setAmount] = useState<number>(0);
   const modalContainer = useRef(null);
-  const [searchQuery, setSearchQuery] = useState("");
-  const debouncedSearch = useDebounce(searchQuery, 300);
-  const [showInput, setShowInput] = useState<boolean>(false);
-  const { data, isError } = useQuery(["data", debouncedSearch], () =>
-    fetchAll(searchQuery),
-  );
-  const [selectedProducts, setSelectedProducts] = useState<Product[]>([]);
   const [step, setStep] = useState<number>(1);
-  const [selectedIngredient, setSelectedIngredient] =
-    useState<NewIngredient | null>(null);
 
-  const { productsByIds, productIds } = data ?? {
-    productIds: [],
-    productsByIds: {},
+  const modalReset = () => {
+    setStep(1);
+    setAmount(0);
+    editSelectedProducts([]);
+    selectIngredient(null);
+    onClose();
   };
-
-  if (isError) {
-    return <div>Error</div>;
-  }
 
   const handleIngredientClick = (ingredient: NewIngredient) => {
     const newSelectedIngredient: NewIngredient = {
       ...ingredient,
       amount: amount,
     };
-    setSelectedIngredients([...selectedIngredients, newSelectedIngredient]);
+    addToSelectedIngredients(newSelectedIngredient);
+    modalReset();
   };
 
   const handleSave = () => {
@@ -94,36 +69,11 @@ const IngredientModal: FC<Props> = ({
         id: Date.now().toString(36),
         amount: amount,
       };
+      addToSelectedIngredients(newIngredient);
       addIngredient(newIngredient);
-      handleIngredientClick(newIngredient);
-      setShowInput(false);
-      setSelectedProducts([]);
-      setSearchQuery("");
-      setName(null);
-      setStep(1);
-      onClose();
-    } else if (name != null && selectedProducts.length === 0) {
-      setError(true);
+      modalReset();
     }
   };
-
-  useEffect(() => {
-    const filteredIngredients = ingredients.filter(
-      (ingredient) =>
-        !selectedIngredients.some(
-          (selected) => selected.id === ingredient.id,
-        ) &&
-        (!selectedIngredient || selectedIngredient.id !== ingredient.id),
-    );
-
-    setCurrentIngredients(filteredIngredients);
-  }, [selectedIngredients, selectedIngredient]);
-
-  useEffect(() => {
-    if (selectedProducts.length > 0) {
-      setError(false);
-    }
-  }, [selectedProducts]);
 
   return (
     <>
@@ -146,48 +96,10 @@ const IngredientModal: FC<Props> = ({
               alignItems={"center"}
             >
               {step === 1 ? (
-                <>
-                  <Heading
-                    fontSize={"24px"}
-                    fontWeight={900}
-                    lineHeight={"32px"}
-                    mb={"24px"}
-                  >
-                    Přidat ingredienci
-                  </Heading>
-                  <Autocomplete
-                    currentIngredients={currentIngredients}
-                    setName={setName}
-                    setSelectedIngredient={setSelectedIngredient}
-                    setShowInput={setShowInput}
-                  />
-                  {showInput ? (
-                    <>
-                      <CreateIngredientInput
-                        searchQuery={searchQuery}
-                        setSearchQuery={setSearchQuery}
-                        productsByIds={productsByIds}
-                        productIds={productIds}
-                        selectedProducts={selectedProducts}
-                        setSelectedProducts={setSelectedProducts}
-                      />
-                      <ChosenProducts
-                        selectedProducts={selectedProducts}
-                        setSelectedProducts={setSelectedProducts}
-                      />
-                    </>
-                  ) : selectedIngredient ? (
-                    <Text>{selectedIngredient.name}</Text>
-                  ) : null}
-                </>
+                <IngredientModalOne setName={setName} />
               ) : (
                 <IngredientModalTwo setAmount={setAmount} />
               )}
-              {error ? (
-                <Text mt={"24px"} color={"red"}>
-                  Vyberte alespoň jeden produkt.
-                </Text>
-              ) : null}
 
               {step === 1 ? (
                 <ButtonGroup
@@ -207,12 +119,7 @@ const IngredientModal: FC<Props> = ({
                     alignItems="center"
                     rounded={"xl"}
                     onClick={() => {
-                      setSearchQuery("");
-                      setSelectedProducts([]);
-                      setSelectedIngredient(null);
-                      setShowInput(false);
-                      setName(null);
-                      onClose();
+                      modalReset();
                     }}
                   >
                     Zrušit
@@ -266,20 +173,7 @@ const IngredientModal: FC<Props> = ({
                     display="flex"
                     alignItems="center"
                     rounded={"xl"}
-                    onClick={
-                      selectedIngredient
-                        ? () => {
-                            setSelectedIngredient(null);
-                            onClose();
-                          }
-                        : () => {
-                            setSearchQuery("");
-                            setSelectedProducts([]);
-                            setShowInput(false);
-                            setName(null);
-                            onClose();
-                          }
-                    }
+                    onClick={() => modalReset()}
                   >
                     Zrušit
                   </Button>
@@ -306,24 +200,29 @@ const IngredientModal: FC<Props> = ({
                     rounded={"xl"}
                     boxShadow={"md"}
                     color={"white"}
+                    isDisabled={amount === 0}
                     _hover={{ bg: "rgb(87, 130, 4)" }}
-                    onClick={
-                      selectedIngredient
-                        ? () => {
-                            handleIngredientClick(selectedIngredient);
-                            onClose();
-                            setSelectedIngredient(null);
-                            setStep(1);
-                          }
-                        : handleSave
-                    }
+                    onClick={() => {
+                      selectedIngredient && amount > 0
+                        ? handleIngredientClick(selectedIngredient)
+                        : handleSave();
+                    }}
                   >
                     Vytvořit
                   </Button>
                 </ButtonGroup>
               )}
-              {/*<Button onClick={() => purgeStorage()}>Purge Storage</Button>*/}
-              {/*bytesInUse: {bytesInUse}*/}
+              {/*<Button*/}
+              {/*  onClick={() => {*/}
+              {/*    resetProducts();*/}
+              {/*    resetIngredient();*/}
+              {/*    resetIngredients();*/}
+              {/*    purgeStorage();*/}
+              {/*  }}*/}
+              {/*>*/}
+              {/*  Purge Storage*/}
+              {/*</Button>*/}
+              {/*/!*bytesInUse: {bytesInUse}*!/*/}
             </Flex>
           </ModalBody>
         </ModalContent>
